+++ util/update-accs.bash
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-for i in "$@"
-do
-  pbsimply-pandoc.rb "$i"
-  pbsimply-accsindex.rb "$i"
-done